requirements.txt
yfinance
pytest
httpx
python-dotenv